requirements.txt
numpy
scipy
matplotlib
python-dotenv
astropy
cartopy
